+++ example/mosmix_forecasts.py
@@ -13,7 +13,7 @@ stations 01001 and 01008 and parameters DD and ww.
 Other MOSMIX variants are also listed and can be
 enabled on demand.
 """
-from wetterdienst.dwd.forecasts import DWDMosmixType, DWDMosmixValues
+from wetterdienst.dwd.forecasts import DWDMosmixStations, DWDMosmixType
 from wetterdienst.dwd.forecasts.metadata.dates import DWDForecastDate
 from wetterdienst.util.cli import setup_logging
 
@@ -21,33 +21,42 @@ from wetterdienst.util.cli import setup_logging
 def mosmix_example():
 
     # A. MOSMIX-L -- Specific stations - each station with own file
-    mosmix = DWDMosmixValues(
-        station_id=["01001", "01008"],
+    request = DWDMosmixStations(
         parameter=["DD", "ww"],
         start_issue=DWDForecastDate.LATEST,  # automatically set if left empty
         mosmix_type=DWDMosmixType.LARGE,
         tidy_data=True,
         humanize_parameters=True,
     )
-    response = next(mosmix.query())
+
+    stations = request.filter(
+        station_id=["01001", "01008"],
+    )
+
+    response = next(stations.values.query())
 
     # meta data enriched with information from metadata_for_forecasts()
-    output_section("Metadata", response.metadata)
-    output_section("Forecasts", response.data)
+    output_section("Metadata", response.stations.df)
+    output_section("Forecasts", response.df)
 
     # B. MOSMIX-S -- All stations - specified stations are extracted.
-    mosmix = DWDMosmixValues(
-        station_id=["01028", "01092"],
+
+    request = DWDMosmixStations(
         parameter=["DD", "ww"],
         start_issue=DWDForecastDate.LATEST,  # automatically set if left empty
         mosmix_type=DWDMosmixType.SMALL,
         tidy_data=True,
         humanize_parameters=True,
     )
-    response = next(mosmix.query())
 
-    output_section("Metadata", response.metadata)
-    output_section("Forecasts", response.data)
+    stations = request.filter(
+        station_id=["01001", "01008"],
+    )
+
+    response = next(stations.values.query())
+
+    output_section("Metadata", response.stations.df)
+    output_section("Forecasts", response.df)
 
 
 def output_section(title, data):  # pragma: no cover