+++ wetterdienst/core/scalar/stations.py
@@ -1,24 +1,23 @@
 # -*- coding: utf-8 -*-
 # Copyright (c) 2018-2021, earthobservations developers.
 # Distributed under the MIT License. See LICENSE for more info.
+import logging
 from abc import abstractmethod
 from datetime import datetime
-from typing import Union, List, Optional
-import logging
+from enum import Enum
+from typing import List, Optional, Tuple, Union
 
 import numpy as np
 import pandas as pd
 import pytz
 
-from wetterdienst import Period, Resolution
 from wetterdienst.core.scalar.core import ScalarCore
 from wetterdienst.core.scalar.result import StationsResult
-from wetterdienst.dwd.metadata.column_names import DWDMetaColumns
 from wetterdienst.metadata.columns import Columns
-from wetterdienst.metadata.period import PeriodType
+from wetterdienst.metadata.period import Period, PeriodType
+from wetterdienst.metadata.resolution import Resolution
 from wetterdienst.util.enumeration import parse_enumeration_from_template
 from wetterdienst.util.geo import Coordinates, derive_nearest_neighbours
-
 
 EARTH_RADIUS_KM = 6371
 
@@ -28,8 +27,15 @@ log = logging.getLogger(__name__)
 class ScalarStationsCore(ScalarCore):
     """ Core for stations information of a source """
 
-    @abstractmethod
     @property
+    @abstractmethod
+    def _parameter_base(self) -> Enum:
+        """parameter base enumeration from which parameters can be parsed e.g.
+        DWDObservationParameter"""
+        pass
+
+    @property
+    @abstractmethod
     def _values(self):
         pass
 
@@ -56,20 +62,62 @@ class ScalarStationsCore(ScalarCore):
         Columns.STATE.value: str,
     }
 
-    def _parse_period(self, period: Period):
+    def _parse_period(self, period: Period) -> Optional[List[Period]]:
         if not period:
             return None
         elif self._period_type == PeriodType.FIXED:
-            return period
+            return [period]
         else:
-            return parse_enumeration_from_template(period, self._period_base, Period)
+            return (
+                pd.Series(period)
+                .apply(
+                    parse_enumeration_from_template, args=(self._period_base, Period)
+                )
+                .sort_values()
+                .tolist()
+            )
+
+    def _parse_parameter(self, parameter: List[Union[str, Enum]]) -> List[Enum]:
+        """
+        Method to parse parameters, either from string or enum. Case independent for
+        strings.
+
+        :param parameter: parameters as strings or enumerations
+        :return: list of parameter enumerations of type self._parameter_base
+        """
+        return (
+            pd.Series(parameter)
+            .apply(parse_enumeration_from_template, args=(self._parameter_base,))
+            .tolist()
+        )
+
+    @staticmethod
+    def _parse_station_id(series: pd.Series) -> pd.Series:
+        """Dedicated method for parsing station ids, by default uses the same method as
+        parse_strings but could be modified by the implementation class"""
+        return series.astype(pd.StringDtype())
+
+    def __eq__(self, other) -> bool:
+        """ Equal method of request object """
+        return (
+            self.parameter == other.parameter
+            and self.resolution == other.resolution
+            and self.period == other.period
+            and self.start_date == other.start_date
+            and self.end_date == other.end_date
+            and self.humanize_parameters == other.humanize_parameters
+            and self.tidy_data == other.tidy_data
+        )
 
     def __init__(
         self,
+        parameter: Tuple[Union[str, Enum]],
         resolution: Resolution,
         period: Period,
-        start_date: Union[None, str, datetime] = None,
-        end_date: Union[None, str, datetime] = None,
+        start_date: Optional[Union[str, datetime]] = None,
+        end_date: Optional[Union[str, datetime]] = None,
+        humanize_parameters: bool = True,
+        tidy_data: bool = True,
     ) -> None:
         """
 
@@ -83,43 +131,53 @@ class ScalarStationsCore(ScalarCore):
             end_date=end_date,
         )
 
-    def all(self) -> StationsResult:
+        self.parameter = self._parse_parameter(parameter)
+        self.humanize_parameters = humanize_parameters
+        self.tidy_data = tidy_data
+
+    def all(self) -> "StationsResult":
         """
         Wraps the _all method and applies date filters.
 
         :return: pandas.DataFrame with the information of different available stations
         """
-        metadata_df = self._all().copy()
+        df = self._all()
 
-        metadata_df = metadata_df.reindex(columns=self._base_columns)
+        df = df.reindex(columns=self._base_columns)
 
-        metadata_df = self._coerce_meta_fields(metadata_df)
+        df = self._coerce_meta_fields(df)
 
-        if self.start_date:
-            metadata_df = metadata_df[
-                metadata_df[DWDMetaColumns.FROM_DATE.value] <= self.start_date
-            ]
-
-        if self.end_date:
-            metadata_df = metadata_df[
-                metadata_df[DWDMetaColumns.TO_DATE.value] >= self.end_date
-            ]
+        # TODO: exchange with more foreceful filtering if user wants
+        # if self.start_date:
+        #     df = df[
+        #         df[Columns.FROM_DATE.value] <= self.start_date
+        #     ]
+        #
+        # if self.end_date:
+        #     df = df[
+        #         df[Columns.TO_DATE.value] >= self.end_date
+        #     ]
 
         # TODO: use StationsResult here
         # TODO: ._filter should stay empty as _all_ stations are returned
 
-        return metadata_df
+        result = StationsResult(self, df.reset_index(drop=True))
 
-    def query(
-            self,
-            station_id: Optional[Union[str, List[str]]] = None,
-            station_name: Optional[Union[str, List[str]]] = None
-    ) -> StationsResult:
+        return result
+
+    def filter(self, station_id: Tuple[str]) -> "StationsResult":
         # TODO: add queries for the given possible parameters
         # TODO: eventually add other parameters from nearby_... and use DataFrame of
         #  them instead
+        df = self.all().df
 
-        pass
+        station_id = self._parse_station_id(pd.Series(station_id))
+
+        df = df[df[Columns.STATION_ID.value].isin(station_id)]
+
+        result = StationsResult(self, df)
+
+        return result
 
     def _coerce_meta_fields(self, df) -> pd.DataFrame:
         """ Method for filed coercion. """
@@ -149,7 +207,7 @@ class ScalarStationsCore(ScalarCore):
         latitude: float,
         longitude: float,
         number: int,
-    ) -> pd.DataFrame:
+    ) -> "StationsResult":
         """
         Wrapper for get_nearby_stations_by_number using the given parameter set. Returns
         nearest stations defined by number.
@@ -164,7 +222,7 @@ class ScalarStationsCore(ScalarCore):
 
         coords = Coordinates(np.array(latitude), np.array(longitude))
 
-        metadata = self.all()
+        metadata = self.all().df
 
         metadata = metadata.reset_index(drop=True)
 
@@ -191,7 +249,7 @@ class ScalarStationsCore(ScalarCore):
             drop=True
         )
 
-        metadata_location[DWDMetaColumns.DISTANCE_TO_LOCATION.value] = distances_km
+        metadata_location[Columns.DISTANCE_TO_LOCATION.value] = distances_km
 
         if metadata_location.empty:
             log.warning(
@@ -202,14 +260,16 @@ class ScalarStationsCore(ScalarCore):
         # TODO: use StationsResult here
         # TODO: add filter for number filtering
 
-        return metadata_location
+        result = StationsResult(self, metadata_location.reset_index(drop=True))
+
+        return result
 
     def nearby_radius(
         self,
         latitude: float,
         longitude: float,
         max_distance_in_km: int,
-    ) -> StationsResult:
+    ) -> "StationsResult":
         """
         Wrapper for get_nearby_stations_by_distance using the given parameter set.
         Returns nearest stations defined by distance (km).
@@ -223,19 +283,22 @@ class ScalarStationsCore(ScalarCore):
         if max_distance_in_km < 0:
             raise ValueError("'max_distance_in_km' has to be at least 0.0.")
 
-        metadata = self.all()
+        metadata = self.all().df
 
-        all_nearby_stations = self.nearby_number(latitude, longitude, metadata.shape[0])
+        all_nearby_stations = self.nearby_number(
+            latitude, longitude, metadata.shape[0]
+        ).df
 
         nearby_stations_in_distance = all_nearby_stations[
-            all_nearby_stations[DWDMetaColumns.DISTANCE_TO_LOCATION.value]
+            all_nearby_stations[Columns.DISTANCE_TO_LOCATION.value]
             <= max_distance_in_km
         ]
 
         # TODO: use StationsResult here
         # TODO: add filter for radius filtering
 
-        return StationsResult(
-            self,
-            nearby_stations_in_distance.reset_index(drop=True)
+        result = StationsResult(
+            self, nearby_stations_in_distance.reset_index(drop=True)
         )
+
+        return result