+++ wetterdienst/dwd/forecasts/api.py
@@ -3,7 +3,6 @@
 # Distributed under the MIT License. See LICENSE for more info.
 import logging
 from datetime import datetime
-from enum import Enum
 from io import StringIO
 from typing import Generator, Optional, Tuple, Union
 from urllib.parse import urljoin
@@ -12,7 +11,9 @@ import pandas as pd
 import requests
 from requests import HTTPError
 
-from wetterdienst.core.scalar import ScalarStationsCore, ScalarValuesCore, ValuesResult
+from wetterdienst.core.scalar.result import StationsResult, ValuesResult
+from wetterdienst.core.scalar.stations import ScalarStationsCore
+from wetterdienst.core.scalar.values import ScalarValuesCore
 from wetterdienst.dwd.forecasts.access import KMLReader
 from wetterdienst.dwd.forecasts.metadata import (
     DWDForecastDate,
@@ -84,196 +85,64 @@ class DWDMosmixValues(ScalarValuesCore):
           https://www.dwd.de/DE/leistungen/opendata/help/schluessel_datenformate/kml/mosmix_elemente_pdf.pdf?__blob=publicationFile&v=2  # noqa:E501,B950
     """
 
-    @property
-    def _source(self) -> Source:
-        return Source.DWD
-
-    @property
-    def _tz(self) -> Timezone:
-        return Timezone.GERMANY
-
-    @property
-    def _data_tz(self) -> Timezone:
-        return Timezone.UTC
-
-    @property
-    def _has_quality(self) -> bool:
-        return False
+    _source = Source.DWD
+    _tz = Timezone.GERMANY
+    _data_tz = Timezone.UTC
+    _has_quality = False
 
     @property
     def _tidy(self) -> bool:
-        return self.tidy_data
+        return self.stations.tidy_data
 
-    @property
-    def _parameter_base(self) -> Enum:
-        return DWDMosmixParameter
-
-    @property
-    def _irregular_parameters(self) -> Tuple[str]:
-        return tuple()
-
-    @property
-    def _integer_parameters(self) -> Tuple[str]:
-        return INTEGER_PARAMETERS
-
-    @property
-    def _string_parameters(self) -> Tuple[str]:
-        return tuple()
+    _irregular_parameters = tuple()
+    _integer_parameters = INTEGER_PARAMETERS
+    _string_parameters = tuple()
 
     _resolution_type = ResolutionType.FIXED
     _resolution_base = None
     _period_type = PeriodType.FIXED
     _period_base = None
 
-    def __init__(
-        self,
-        station_id: Tuple[str],
-        mosmix_type: Union[str, DWDMosmixType],
-        parameter: Optional[Tuple[Union[str, DWDMosmixParameter]]] = None,
-        start_issue: Optional[
-            Union[str, datetime, DWDForecastDate]
-        ] = DWDForecastDate.LATEST,
-        end_issue: Optional[Union[str, datetime]] = None,
-        start_date: Optional[Union[str, datetime]] = None,
-        end_date: Optional[Union[str, datetime]] = None,
-        humanize_parameters: bool = False,
-        tidy_data: bool = True,
-    ) -> None:
-        """
+    def __init__(self, stations: StationsResult) -> None:
+        """"""
+        super(DWDMosmixValues, self).__init__(stations=stations)
 
-        :param station_id: station ids which are being queried from the MOSMIX foreacst
-        :param mosmix_type: type of forecast, either small (MOSMIX-S) or large
-                            (MOSMIX-L), as string or enumeration
-        :param parameter: optional parameters for which the forecasts are filtered
-        :param start_issue: start date of the MOSMIX forecast, can be used in
-                            combination with end_issue to query multiple MOSMIX
-                            forecasts, or instead used with enumeration to only query
-                            LATEST MOSMIX forecast
-        :param end_issue: end issue of MOSMIX forecast, can be used to query multiple
-                          MOSMIX forecasts available on the server
-        :param start_date: start date to limit the returned data to specified datetimes
-        :param end_date: end date to limit the returned data to specified datetimes
-        :param humanize_parameters: boolean if parameters shall be renamed to human
-                                    readable names
-        :param tidy_data: boolean if pandas.DataFrame shall be tidied and
-                            values put in rows
-        """
-        # Use all parameters if none are given
-        parameter = parameter or [*self._parameter_base]
-
-        super(DWDMosmixValues, self).__init__(
-            station_id=station_id,
-            parameter=parameter,
-            resolution=Resolution.HOURLY,
-            period=Period.FUTURE,
-            start_date=start_date,
-            end_date=end_date,
-            humanize_parameters=humanize_parameters,
-            tidy_data=tidy_data,
+        self.kml = KMLReader(
+            station_ids=self.stations.station_id.tolist(),
+            parameters=[parameter for parameter in self.stations.parameter],
         )
-        self.mosmix_type = parse_enumeration_from_template(mosmix_type, DWDMosmixType)
-
-        # Parse issue date if not set to fixed "latest" string
-        if start_issue is DWDForecastDate.LATEST and end_issue:
-            log.info(
-                "end_issue will be ignored as 'latest' was selected for issue date"
-            )
-
-        if start_issue is not DWDForecastDate.LATEST:
-            if not start_issue and not end_issue:
-                start_issue = DWDForecastDate.LATEST
-            elif not end_issue:
-                end_issue = start_issue
-            elif not start_issue:
-                start_issue = end_issue
-
-            start_issue = pd.to_datetime(start_issue, infer_datetime_format=True).floor(
-                "1H"
-            )
-            end_issue = pd.to_datetime(end_issue, infer_datetime_format=True).floor(
-                "1H"
-            )
-
-            # Shift start date and end date to 3, 9, 15, 21 hour format
-            if mosmix_type == DWDMosmixType.LARGE:
-                start_issue = self.adjust_datetime(start_issue)
-                end_issue = self.adjust_datetime(end_issue)
-
-        self.start_issue = start_issue
-        self.end_issue = end_issue
-        self.humanize_parameters = humanize_parameters
-        self.tidy_data = tidy_data
-
-        # TODO: this should be replaced by the freq property in the main class
-        if self.mosmix_type == DWDMosmixType.SMALL:
-            self.freq = "1H"  # short forecasts released every hour
-        else:
-            self.freq = "6H"
-
-        self.kml = KMLReader(station_ids=self.station_ids, parameters=self.parameters)
 
     # TODO: add __eq__ and __str__
 
     @property
     def metadata(self) -> pd.DataFrame:
         """ Wrapper for forecast metadata """
-        return DWDMosmixStations().all()
-
-    @staticmethod
-    def adjust_datetime(datetime_: datetime) -> datetime:
-        """
-        Adjust datetime to MOSMIX release frequency, which is required for MOSMIX-L
-        that is only released very 6 hours (3, 9, 15, 21). Datetime is floored
-        to closest release time e.g. if hour is 14, it will be rounded to 9
-
-        Args:
-            datetime_: datetime that is adjusted
-
-        Returns:
-            adjusted datetime with floored hour
-        """
-        regular_date = datetime_ + pd.offsets.DateOffset(hour=3)
-
-        if regular_date > datetime_:
-            regular_date -= pd.Timedelta(hours=6)
-
-        delta_hours = (datetime_.hour - regular_date.hour) % 6
-
-        datetime_adjusted = datetime_ - pd.Timedelta(hours=delta_hours)
-
-        return datetime_adjusted
+        return self.stations.df
 
     def query(self) -> Generator[ValuesResult, None, None]:
         """Replace collect data method as all information is read once from kml file"""
-        for metadata_df, forecast_df in self._collect_station_parameter():
+        for forecast_df in self._collect_station_parameter():
             forecast_df = self._coerce_meta_fields(forecast_df)
             forecast_df = self._coerce_parameter_types(forecast_df)
 
-            if self.humanize_parameters:
+            if self.stations.humanize_parameters:
                 forecast_df = self._humanize(forecast_df)
 
-            # Complement metadata
-            station_id = forecast_df[DWDMetaColumns.STATION_ID.value].iloc[0]
-
-            station_metadata = self.metadata[
-                self.metadata[DWDMetaColumns.STATION_ID.value] == station_id
-            ].reset_index(drop=True)
-
-            metadata_df = metadata_df.rename(columns=str.upper).reset_index(drop=True)
-
-            metadata_df = metadata_df.join(station_metadata)
-
-            result = ValuesResult(metadata_df, forecast_df)
+            result = ValuesResult(stations=self.stations, df=forecast_df)
 
             yield result
 
-    def _collect_station_parameter(self) -> Generator[ValuesResult, None, None]:
+    def _collect_station_parameter(self) -> Generator[pd.DataFrame, None, None]:
         """Wrapper of read_mosmix to collect forecast data (either latest or for
         defined dates)"""
-        if self.start_issue == DWDForecastDate.LATEST:
-            yield from self.read_mosmix(self.start_issue)
+        if self.stations.start_issue == DWDForecastDate.LATEST:
+            yield from self.read_mosmix(self.stations.stations.start_issue)
         else:
-            for date in pd.date_range(self.start_issue, self.end_issue, freq=self.freq):
+            for date in pd.date_range(
+                self.stations.stations.start_issue,
+                self.stations.stations.end_issue,
+                freq=self.stations.frequency.value,
+            ):
                 try:
                     yield from self.read_mosmix(date)
                 except IndexError as e:
@@ -288,7 +157,7 @@ class DWDMosmixValues(ScalarValuesCore):
         :param date: datetime or enumeration for latest MOSMIX forecast
         :return: DWDMosmixResult with gathered information
         """
-        for df_metadata, df_forecast in self._read_mosmix(date):
+        for df_forecast in self._read_mosmix(date):
             df_forecast = df_forecast.rename(
                 columns={
                     "station_id": DWDMetaColumns.STATION_ID.value,
@@ -296,7 +165,7 @@ class DWDMosmixValues(ScalarValuesCore):
                 }
             )
 
-            if self.tidy_data:
+            if self.stations.tidy_data:
                 df_forecast = df_forecast.melt(
                     id_vars=[
                         DWDMetaColumns.STATION_ID.value,
@@ -306,14 +175,14 @@ class DWDMosmixValues(ScalarValuesCore):
                     value_name=DWDMetaColumns.VALUE.value,
                 )
 
-            yield df_metadata, df_forecast
+            yield df_forecast
 
     def _read_mosmix(
         self, date: Union[DWDForecastDate, datetime]
-    ) -> Generator[Tuple[pd.DataFrame, pd.DataFrame], None, None]:
+    ) -> Generator[pd.DataFrame, None, None]:
         """Wrapper that either calls read_mosmix_s or read_mosmix_l depending on
         defined period type"""
-        if self.mosmix_type == DWDMosmixType.SMALL:
+        if self.stations.stations.mosmix_type == DWDMosmixType.SMALL:
             yield from self.read_mosmix_small(date)
         else:
             yield from self.read_mosmix_large(date)
@@ -330,7 +199,7 @@ class DWDMosmixValues(ScalarValuesCore):
         self.kml.read(file_url)
 
         for forecast in self.kml.get_forecasts():
-            yield self.kml.get_metadata(), forecast
+            yield forecast
 
     def read_mosmix_large(
         self, date: Union[DWDForecastDate, datetime]
@@ -339,7 +208,7 @@ class DWDMosmixValues(ScalarValuesCore):
         forecast per file."""
         url = urljoin(DWD_SERVER, DWD_MOSMIX_L_SINGLE_PATH)
 
-        for station_id in self.station_ids:
+        for station_id in self.stations.station_id:
             station_url = f"{url}{station_id}/kml"
 
             try:
@@ -350,7 +219,7 @@ class DWDMosmixValues(ScalarValuesCore):
 
             self.kml.read(file_url)
 
-            yield self.kml.get_metadata(), next(self.kml.get_forecasts())
+            yield next(self.kml.get_forecasts())
 
     @staticmethod
     def get_url_for_date(url: str, date: Union[datetime, DWDForecastDate]) -> str:
@@ -397,13 +266,15 @@ class DWDMosmixValues(ScalarValuesCore):
 class DWDMosmixStations(ScalarStationsCore):
     """ Implementation of sites for MOSMIX forecast sites """
 
-    @property
-    def _source(self) -> Source:
-        return Source.DWD
+    _source = Source.DWD
+    _tz = Timezone.GERMANY
+    _parameter_base = DWDMosmixParameter
+    _values = DWDMosmixValues
 
-    @property
-    def _tz(self) -> Timezone:
-        return Timezone.GERMANY
+    _resolution_type = ResolutionType.FIXED
+    _resolution_base = None
+    _period_type = PeriodType.FIXED
+    _period_base = None
 
     _base_columns = [
         Columns.STATION_ID.value,
@@ -417,20 +288,101 @@ class DWDMosmixStations(ScalarStationsCore):
         Columns.STATE.value,
     ]
 
-    _resolution_type = ResolutionType.FIXED
-    _resolution_base = None
-    _period_type = PeriodType.FIXED
-    _period_base = None
+    @staticmethod
+    def adjust_datetime(datetime_: datetime) -> datetime:
+        """
+        Adjust datetime to MOSMIX release frequency, which is required for MOSMIX-L
+        that is only released very 6 hours (3, 9, 15, 21). Datetime is floored
+        to closest release time e.g. if hour is 14, it will be rounded to 9
 
-    def __init__(self) -> None:
+        Args:
+            datetime_: datetime that is adjusted
+
+        Returns:
+            adjusted datetime with floored hour
+        """
+        regular_date = datetime_ + pd.offsets.DateOffset(hour=3)
+
+        if regular_date > datetime_:
+            regular_date -= pd.Timedelta(hours=6)
+
+        delta_hours = (datetime_.hour - regular_date.hour) % 6
+
+        datetime_adjusted = datetime_ - pd.Timedelta(hours=delta_hours)
+
+        return datetime_adjusted
+
+    def __init__(
+        self,
+        mosmix_type: Union[str, DWDMosmixType],
+        parameter: Optional[Tuple[Union[str, DWDMosmixParameter]]] = None,
+        start_issue: Optional[
+            Union[str, datetime, DWDForecastDate]
+        ] = DWDForecastDate.LATEST,
+        end_issue: Optional[Union[str, datetime]] = None,
+        humanize_parameters: bool = True,
+        tidy_data: bool = True,
+    ) -> None:
+        parameter = parameter or [*self._parameter_base]
+
         super().__init__(
+            parameter=parameter,
             start_date=None,
             end_date=None,
             resolution=Resolution.HOURLY,
             period=Period.FUTURE,
         )
 
-    def _all(self):
+        self.mosmix_type = parse_enumeration_from_template(mosmix_type, DWDMosmixType)
+
+        # Parse issue date if not set to fixed "latest" string
+        if start_issue is DWDForecastDate.LATEST and end_issue:
+            log.info(
+                "end_issue will be ignored as 'latest' was selected for issue date"
+            )
+
+        if start_issue is not DWDForecastDate.LATEST:
+            if not start_issue and not end_issue:
+                start_issue = DWDForecastDate.LATEST
+            elif not end_issue:
+                end_issue = start_issue
+            elif not start_issue:
+                start_issue = end_issue
+
+            start_issue = pd.to_datetime(start_issue, infer_datetime_format=True).floor(
+                "1H"
+            )
+            end_issue = pd.to_datetime(end_issue, infer_datetime_format=True).floor(
+                "1H"
+            )
+
+            # Shift start date and end date to 3, 9, 15, 21 hour format
+            if mosmix_type == DWDMosmixType.LARGE:
+                start_issue = self.adjust_datetime(start_issue)
+                end_issue = self.adjust_datetime(end_issue)
+
+        # TODO: this should be replaced by the freq property in the main class
+        if self.mosmix_type == DWDMosmixType.SMALL:
+            self.resolution = Resolution.HOURLY
+        else:
+            self.resolution = Resolution.HOUR_6
+
+        self.start_issue = start_issue
+        self.end_issue = end_issue
+        self.humanize_parameters = humanize_parameters
+        self.tidy_data = tidy_data
+
+    @property
+    def issue_start(self):
+        """ Required for typing """
+        return self.issue_start
+
+    @property
+    def issue_end(self):
+        """ Required for typing """
+        return self.issue_end
+
+    def _all(self) -> pd.DataFrame:
         """ Create meta data DataFrame from available station list """
         # TODO: Cache payload with FSSPEC
         payload = requests.get(MOSMIX_STATION_LIST, headers={"User-Agent": ""})
@@ -477,4 +429,4 @@ class DWDMosmixStations(ScalarStationsCore):
 
         df = df.reindex(columns=MOSMIX_METADATA_COLUMNS)
 
-        return df
+        return df.copy()