+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ Changelog
 Development
 ***********
 
-...
+- Add StationsResult and ValuesResult to allow for new workflow and connect stations and
+  values request
+- Add accessor .values to Stations class to get straight to values for a request
 
 0.14.1 (21.02.2021)
 *******************